+++ server/camcops_server/tasks/__init__.py
@@ -116,6 +116,7 @@ from camcops_server.tasks.npiq import NpiQ
 from camcops_server.tasks.ors import Ors
 
 from camcops_server.tasks.panss import Panss
+from camcops_server.tasks.pbq import Pbq
 from camcops_server.tasks.pcl5 import Pcl5
 from camcops_server.tasks.pcl import PclC, PclM, PclS
 from camcops_server.tasks.pdss import Pdss